Put all layers in one jit. This creates less kernels and is slightly faster. Run on RTX 3090

I also fixed loading in the fp16 weights such that they are actually used properly and reduce GPU mem.

`JIT=1 python examples/gpt2.py --timing --count 30 --seed 0 --model_size gpt2`
Before
```
JIT captured 4 kernels with 2 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 20 kernels with 3 inputs
JIT captured 8 kernels with 1 inputs

ran model in 4.82 ms
total 6.72 ms
ran model in 4.80 ms
total 6.70 ms
ran model in 4.80 ms
total 6.70 ms
```
After:
```
JIT captured 240 kernels with 26 inputs

ran model in 4.69 ms
total 6.65 ms
ran model in 4.42 ms
total 6.39 ms
ran model in 4.47 ms
total 6.45 ms
```

With gpt2-medium
Before:
```
ran model in 8.98 ms
total 13.81 ms

ran model in 9.00 ms
total 13.82 ms
```

After:
```
ran model in 8.34 ms
total 13.64 ms

ran model in 8.33 ms
total 13.68 ms
```